+++ packages/constructor/src/constructors/TextDescription/index.ts
@@ -1,0 +1,5 @@
+export * from './TextDescription'
+export * from './TextDescriptionDesign'
+export * from './props'
+export * from './types'
+export * from './basicTypes'