+++ packages/d1/src/components/Ui/TextDescription/wiki.ts
@@ -1,0 +1,13 @@
+import { WikiStorybook } from '@dxtmisha/wiki'
+import { wiki, wikiDescriptions } from '@dxtmisha/wiki/media'
+
+import { TextDescriptionWikiData } from './wikiData'
+
+export const TextDescriptionWikiStorybook = new WikiStorybook(
+  TextDescriptionWikiData.component,
+  TextDescriptionWikiData.props,
+  TextDescriptionWikiData.defaults,
+  TextDescriptionWikiData.wikiDesign,
+  wiki,
+  wikiDescriptions
+)