+++ packages/d1/src/components/Ui/TextDescription/wikiData.ts
@@ -1,0 +1,32 @@
+import { type StorybookProps, type StorybookSlots, type WikiDataItem } from '@dxtmisha/wiki'
+
+import { wikiD1 } from '../../../wiki/wiki'
+import { defaults } from './props'
+
+const propsNames: StorybookProps = [
+  // :propsList [!] System label / Системная метка
+  { name: 'description', type: 'string | number | undefined' },
+  { name: 'descriptionId', type: 'string | undefined' },
+  { name: 'size', type: 'string', option: ['xs', 'sm', 'md'] },
+  { name: 'tag', type: 'string | undefined' }
+  // :propsList [!] System label / Системная метка
+]
+
+const slotsNames: StorybookSlots = [
+  // :slotsList [!] System label / Системная метка
+  // :slotsList [!] System label / Системная метка
+]
+
+const eventsNames: StorybookSlots = [
+  // :eventsList [!] System label / Системная метка
+  // :eventsList [!] System label / Системная метка
+]
+
+export const TextDescriptionWikiData: WikiDataItem = {
+  component: 'TextDescription',
+  props: propsNames,
+  slots: slotsNames,
+  events: eventsNames,
+  defaults,
+  wikiDesign: wikiD1
+}